+++ logger.go
@@ -1,64 +1,50 @@
 package logger
 
 import (
-	"os"
-	"time"
-
-	strutils "github.com/iolave/go-logger/pkg/str_utils"
+	"context"
 )
 
-type Logger struct {
-	name string
-}
+type Logger interface {
+	// Trace logs a message with the LEVEL_TRACE level.
+	Trace(ctx context.Context, msg string)
 
-func New(name string) *Logger {
-	return &Logger{
-		name: name,
-	}
-}
+	// TraceWithData logs a message with the LEVEL_TRACE level
+	// and the given data.
+	TraceWithData(ctx context.Context, msg string, data map[string]any)
 
-func (log Logger) Info(msg string, customData map[string]interface{}) {
-	entry := log.buildLogEntry(LOG_LEVEL_INFO, msg, customData)
-	entry.print()
-}
+	// Debug logs a message with the LEVEL_DEBUG level.
+	Debug(ctx context.Context, msg string)
 
-func (log Logger) Warn(msg string, customData map[string]interface{}) {
-	entry := log.buildLogEntry(LOG_LEVEL_WARN, msg, customData)
-	entry.print()
-}
+	// DebugWithData logs a message with the LEVEL_DEBUG level
+	// and the given data.
+	DebugWithData(ctx context.Context, msg string, data map[string]any)
 
-func (log Logger) Error(msg string, customData map[string]interface{}) {
-	entry := log.buildLogEntry(LOG_LEVEL_ERROR, msg, customData)
-	entry.print()
-}
+	// Info logs a message with the LEVEL_INFO level.
+	Info(ctx context.Context, msg string)
 
-func (log Logger) Debug(msg string, customData map[string]interface{}) {
-	entry := log.buildLogEntry(LOG_LEVEL_DEBUG, msg, customData)
-	entry.print()
-}
+	// InfoWithData logs a message with the LEVEL_INFO level
+	// and the given data.
+	InfoWithData(ctx context.Context, msg string, data map[string]any)
 
-func (log Logger) Fatal(msg string, customData map[string]interface{}) {
-	entry := log.buildLogEntry(LOG_LEVEL_FATAL, msg, customData)
-	entry.print()
-	os.Exit(1)
-}
+	// Warn logs a message with the LEVEL_WARN level.
+	Warn(ctx context.Context, msg string, err error)
 
-func (log Logger) buildLogEntry(level LogLevel, msg string, customData map[string]interface{}) LogEntry {
-	entry := new(LogEntry)
+	// WarnWithData logs a message with the LEVEL_WARN level
+	// and the given data.
+	WarnWithData(ctx context.Context, msg string, err error, data map[string]any)
 
-	hostname, _ := os.Hostname()
+	// Error logs a message with the LEVEL_ERROR level.
+	Error(ctx context.Context, msg string, err error)
 
-	// Setting base log entry fields
-	entry.Name = log.name
-	entry.Level = level
-	entry.Time = int(time.Now().Unix())
-	entry.Pid = os.Getpid()
-	entry.Hostname = hostname
-	entry.SchemaVersion = "v1.0.0" // TODO: Add schema definition in README.md
+	// ErrorWithData logs a message with the LEVEL_ERROR level
+	// and the given data.
+	ErrorWithData(ctx context.Context, msg string, err error, data map[string]any)
 
-	// Setting user log info
-	entry.Msg = strutils.ToSnakeCase(msg)
-	entry.CustomData = customData
+	// Fatal logs a message with the LEVEL_FATAL level and it's followed by a
+	// call to os.Exit(1).
+	Fatal(ctx context.Context, msg string, err error)
 
-	return *entry
+	// FatalWithData logs a message with the LEVEL_FATAL level
+	// and the given data and it's followed by a call to os.Exit(1).
+	FatalWithData(ctx context.Context, msg string, err error, data map[string]any)
 }